Update logic around link colors
With this new approach, users can use a core/Link directly and still get
updated link styles when using a ui/Theme. This is now done through CSS
style sheets instead of by manually setting attributes on each and every
<a> tag rendered by Compose.

Note that this API is still under consideration, and we may end up
adding a ui/Link back in, similar to how Chakra-UI works, in case the
nekt theming library starts to use a different set of standards when
naming input parameters.