Remove unecessary dependency
The UI module briefly depended on the Core module, but I was able to
remove that coupling in a previous commit.

I need to think this through a bit more, but we should probably strive
to keep the UI module totally separate from Core, as that might make it
a bit more flexible if we ever end up in an ecosystem like TypeScript+
React, and want to have a UI layer that can potentially work with other
solutions.

Either way, it's not necessary right now, so let's just remove it!