Convert "Link" to "PageLink"
"Link" came from next.js, but I always found the component a little
confusing, because it both wants to be and not be an <a/> tag at the
same time. In practice, they recommend:

<Link href="/about"><a>About</a></Link>

which seems a little redundant.

The final API is still forming in my mind, but I'm starting to adopt
 the following pattern:

```
fun AttrsBuilder.makeComponent(params: Params)
```

as the general solution, with a specific

```
@Composable
fun Component(params: Params)
```

as an opinionated shortcut.

So, for example, for PageLink, you can do:

```
A({
  attrs = { makePageLink(params) }
}
```

OR

```
Div({
  attrs = { makePageLink(params) }
}
```

OR

```
PageLink(params)
// Equivalent to A({ attrs = ... }) version
```