Add (temporary?) Color class that supports alpha
For the time being, web compose multiplatform does not support alpha.
This is likely a temporary situation, see also:
https://github.com/JetBrains/compose-jb/issues/943

Until that is fixed, we'll introduce our own color class for now. There
are "toJbColor" and "toSilkColor" extension methods in case we need to
convert between the two on the fly.