Simplfy WebModifier logic
WebModifiers now only contain a single AttrsBuilder, instead of a list
of StyleBuilders and AttrBuilders, which was kind of wasteful as usually
most modifiers only had one or the other but not both. This works fine
because we rely on the "Modifier.then" and "Modifier.fold" methods to
handle the list behavior in a more general, higher-level manner.