Removed git stuff for now
It doesn't feel great without the ability to create an initial commit,
and it's hard to explain to users that it's lack is due to JGit's
missing feature set (or maybe I'm just not finding the right way to use
it).

At least, now the code lives in the history, so even though we're
removing it for now, we can come back later and try readding again in
the future.