Handle CLI run better if server is dead
Previously, we assumed that if a server/state.yaml file was found that
the server was alive, but actually, that's not a guarantee. Now, we
additionally check the server's PID to make sure it's ACTUALLY alive.

Previous, if you started a server and then kill -9'd it, `kobweb run`
would fail to ever startup.