Re-revert decision on "color" params
The realization I had is that html, which allows using "color"
anywhere (because it gets inherited by children) isn't how Compose
works (which uses the theme if nothing is specified). So in Silk,
if you want to set a foreground color, it has to be explicitly as a
parameter, instead of as a Modifier property.

Modifier.backgroundColor is still allowed, however (as it, too,
exists in the Android Compose API)