Update "publish" logic to only push JS artifacts
Compose-aware artifacts at the moment need to be defined within a
multiplatform environment, because the plugin applies the Compose
compiler plugin for you automatically. However, this was generating
extra artifacts in our maven repository that we didn't need.

Now, we disable those tasks, which reduce task execution time,
simplifies our artifact collection in the repo, and means we're
using less disk space to boot.