Simplify Page annotation for now
Before, the annotation provided additional ways to change the route
associated with that page, but they aren't ready yet, so for now we
just remove it. We can always dig the code back from the code
history if we need to.

This cleanup is nice to do before we announce Kobweb, since we
shouldn't expose an API surface that isn't supported.