Rename "kobweb" artifact to "kobweb-core"
This makes the name more parallel to the its sibling artifacts but,
perhaps more important, seems to fix a bug in the IDE that gets
confused by the name (perhaps the existance of other "kobweb-xyz"
artifacts makes it think that "kobweb" is a container rather than an
artifact in its own right).

We're following, essentially, the same pattern here used by the
Kotlin coroutines project. This also allows the artifact name to
match the name we use for it in the versions.toml file.