Remove generics from ComponentStyle API
... renaming it to ComponentModifiers for now, in case we decide to
revert back to the original API after an upstream Compose compiler
bug is fixed.

The API is a bit simpler now and easier to read; however, we have
to accept a "data: Any?" parameter which is often null, which is
a bit awkward.

We'll let this new API settle a bit as an experiment for now and
revisit it before 1.0