Totally revamp ComponentStyle and SilkTheme approach
SilkTheme is no longer composable. Instead, it is set once and baked in at
initialization time. We are now more aligned with the way the web naturally
works.

For example, before, Button styling looked like this:

```
val ButtonKey = ComponentKey("silk-button")
object ButtonModifier : ComponentModifier {
  @Composable
  override fun toModifier(data: Any?): Modifier {
    val state = data as ButtonState
    val modifier =
      when (state) {
        ButtonState.DEFAULT -> SilkTheme.palette.primary
        ButtonState.HOVER -> SilkTheme.palette.primary.shifted()
        ButtonState.PRESSED -> SilkTheme.palette.primary.shifted().shifted()
      }.let { color -> Modifier.background(color) }

      return modifier.clip(Rect(4.px))
  }
}
```

which additionally required we add a bunch of logic to keep our own copy of the
button's "normal", "hover", and "pressed" states (not included in this README,
but that was another ~15 lines of code).

Now, it looks like this:

```
val ButtonStyle = ComponentStyle("silk-button") { colorMode ->
  val buttonColors = SilkTheme.palettes[colorMode].button
  base = Modifier.background(buttonColors.default).clip(Rect(4.px))
  hover = Modifier.background(buttonColors.hover)
  active = Modifier.background(buttonColors.pressed)
}
```

Way simpler, and with this new approach, we now support link styling as
well, which wasn't possible before (since whether a link is visited or not
is hidden from the developer).

For a deeper dive into the reason behind this change, see:
https://docs.google.com/document/d/151FTX1N4-s4SQfYTvN2akvk8tNgM5sDaA8Nao9I8_j4/