Begin refactoring all our modifiers into a subpackage
I'd like to potentially encourage users to use a * import on all the
modifiers extensions we have (in much the same way that Web Compose has
a bunch of extension methods for css).

Even if a user doesn't want to do this, it's still expected that the
number of methods that we're going to add here will explode over the
next few weeks, so best to put it in its own subpackage now before it
gets even harder.