Change ComponentModifiers style from property to function
For example, what used to be:

```
ComponentStyle(...) {
  active = Modifier.color(Colors.Red)
}
```

is now

```
ComponentStyle(...) {
  active {
    Modifier.color(Colors.Red)
  }
}
```

On the surface of it, these feels needlessly verbose, but it future-proofs
us to more complex selectors in the future, such as:

```
ComponentStyle(...) {
  nthChild(4) {
    Modifier.color(Colors.Green)
  }
}
```

This also has the added bonus of communicating that setting modifiers is a
one-way operation. You can set, but you can't get.