Update API route behavior to require setting body or status
Before, this would return 200 (success):

```
@Api
fun doNothing(ctx: ApiContext) {
  // Do nothing!
}
```

but now we consider unhandled requests a 400 (failure).

You either need to set the status code AND/or the response body.

```
@Api
fun doSomething(ctx: ApiContext) {
  if (ctx.req.method == GET) {
    ctx.res.setBodyText("HELLO") // Also sets status to 200
  }
}
```

The above would return a 200 for a GET request and a 400 for any other
sort of request.

Fixes #44