Pass "no-sandbox" arg into Chrome when doing an export
The reason, also added as a comment inside the code:

Normally "no-sandbox" is NOT recommended for security reasons.
However, this option is necessary when this task runs as root,
as Chrome complains otherwise, but this scenario is common in
containers. (It's expected that `kobweb export` will often get
run as part of a container image built in the Cloud). Since
the lifetime of the browser is short, and it exists only for
doing these exports of our own code before getting shut down
again, the fact we are disabling the sandbox here is not a
concern (as far as I can think through).