Split up web-compose-ext into two modules
... one part for the stuff that's simply missing from web compose,
and one part for the additions we're adding (Modifier, Box, etc.)
as part of Kobweb. The former is still just "web-compose-ext" while
the latter is called "kobweb-compose"

While this may look like busiwork now, the separation is useful
because after a followup change, "kobweb-compose" will require
initialization to use (styles will be moving to a stylesheet),
while the original web-compose-ext stuff won't. This is a useful
distinction because it is hoped that one day web-compose-ext will
go away if its changes get forked upstream.