Add "-self" suffix for flex layout modifications
Originally, it was assumed that the stylesheet would be enough
to disambiguate between the two cases, but when, for example, a
Column was the child of another Column, you actually needed two
separate classes for the inner one - one to represent how it should
be laid out, and one to represent how its children should be laid
out (since it is, itself, the child of a parent Column).

Basically, before this fix, columns within columns (or rows within
rows, or boxes within boxes) would cause trouble.