Nit: Move canvas docs to the public methods
When I tried using a canvas in a demo project, it was hard to know
what the fields did because the docs didn't autocomplete.

I originally defined them against the private implementation in order
to avoid duplicating my header docs, but in practice, it resulted in
a bad user experience.