Server routing: Upgrade IOException catch to general Exception
I'm not 100% sure yet, but I think it's this handler
not handling an exception that causes ktor to abort
when trying to shut down. Instead of trying to figure
this out one exception at a time, it seems easier just to
catch all. Especially the last exception I saw was a
coroutine exception spawned by the system, but I don't want
to add a dependency on the library just so I could reference
that exception (JobCancellationException)