Ignore empty component styles
This can be useful as a way to avoid generating a bunch of intermediate
DOM noise for styles that don't actually do anything. Previously, we
would always tag a widget with a classname, even if the class style didn't
do anything.

I'm still not 100% sure this is what users expect, but my motivation behind
this change is to make the experience more consistent with inline styles.

That is, if you try to use inline styles, like this:

```
SomeWidget(Modifier) { ... }
```

then the DOM tree will just ignore it.

Before, if you did this:

```
val WidgetStyle = ComponentStyle("widget-style") {
  base { Modifier }
}

SomeWidget(WidgetStyle.toModifier())
```

suddenly you're generating a bunch of DOM noise for an
empty CSS class.

If you were writing your code by hand, then you probably
would include the class explicitly, even if it was empty,
because maybe you would change the stylesheet later. But with
Kobweb, you would change the code, not the stylesheet.

If you ever wanted to explicitly add a class-name instead,
you could always do:

```
SomeWidget(Modifier.className("widget-style"))
```

By making this implementation choice, I open up the ability for
widget authors to specify an empty style themselves that users
can replace later, and only if they did so, would that particular
codebase end up with a more verbose DOM tree.