Add support for static site layouts
This should make it possible for users to export their site in
a format easy to deploy on a static hosting site, e.g. Netlify

This disables some of the features that make Kobweb powerful
(e.g. dynamic routes and server API routes), but for many use
cases, e.g. portfolios or blog sites, this is fine.

Bug #22