Fix bug with detecting duplicate style / variant logic
Due to the way Kobweb works, there might be a couple of cases
where the same style or variant is re-registered, which isn't
amazing but shouldn't cause a forced crash. The only case that
should break is if a user tries to register a *different* style
or variant with the same name, because that indicates potential
user error (e.g. a copy / paste bug)

Also: update the helloworld sample with code that would hvae
triggered this.