Fix displayIf behavior
Previously, displayIf would not play nicely with any element that
itself set its own display value, such as Row (which set its display
value to Flex), because it would override it unecessarily, deleting
that display value even if in theory the size of the element was
large enough.

By switching the way it handles its media query, we can avoid this
case and use the DisplayIf modifiers more confidently.