Remove special key "root" from MarkdownContext
This avoids having to deal with a nasty double-escaping case
with the value in root, e.g.

```
root: PageLayout(title = "With \"Inner\" Quotes")
```

which would then try to get serialized inside a value
String:

"root" to "PageLayout(title = \"With \\"Inner\\" Quotes")"

We may still have to deal with escaped strings in the future,
but still, as there shouldn't be a usecase to access "root"
inside a user's project at the moment, we can at least sidestep
this until someone explicitly requests the feature with a strong
usecase.