Fix kobwebGenSite... dependencies
A previous refactor split up "kobwebGenSiteIndex" and
"kobwebGenSiteSources", which should have been safe to do
(famous last words). But what actually happened is some users
could get into a state (by switching branches?) where their code
would get recompiled but the index.html file would not. This
normally isn't a problem -- the "index.html" file only references
code indirectly by its filename -- but what was happening was that
the index file would have gotten built with a "release" target
while the code got rebuilt using a "debug" target. This would
result in the code referencing an HTML element that was supposed
to be found in the index.html file that wasn't there.

This didn't happen before because switching branches before caused
a source code change, which in turn also caused the index.html file
to get rebuilt due to the monolothic task.

Instead of wiring "kobwebGenSiteIndex" to forcefully depend on the
output of "kobwebGenSiteSource", instead, I noticed that the real
root problem was that the build target had changed but the task
didn't notice. I thought Gradle would have handled this automatically
(since the build target was passed in as a constructor argument).
Anyway, for now, we explicitly wire up the build target as an
input dependency. Hopefully this resolves the issue once and for
all.