Rework ref scope logic to a version that takes keys
The first pass of this feature only allowed users to register a
single ref listener which, furthermore, didn't take any keys which
could be used to invalidate the DOM element.

After talking with users on my Discord, however, it turns out that
key invalidation was more important than I thought, for example a Row
widget that should invalidate if the UI has scrolled.

At this point, all Silk widgets take an optional `ref` parameter,
which, if set, can be used to add listeners to the underlying raw
elements that back the composable. There are three ways to construct
it: `ref { ... }`, `disposableRef { ... }` and `refScope`