Remove @Deprecated markers from JB legacy alignment Modifiers
Even though users should prefer the Kobweb versions, the experience
around importing the Kobweb version over the JB version isn't great.
The IDE simply doesn't offer a great way to import just the path
you need, which means that users will see a deprecated method without
an easy way to fix it besides manually adding the import themselves.

The fix that the Kobweb version of the alignment modifiers brings is
so minor that probably less than 1% of all users will actually care
about it, if I had to guess.

So for those reasons together, we just leave with the confusing fact
that there are two versions of AlignSelf/Items/Content, and leave it
up to the docs to help users navigate the choices.