Update `kobweb create` logic around finding Kobweb root
Soon, we'll release a multimodule template, which doesn't have its
Kobweb folder in the root, but rather in a subfolder.

At the end of running `kobweb create`, you see a message like
this:

```
We suggest typing:

$ cd site
$ kobweb run
```
which was actually semi-hardcoded.

But now, after this change, someone can put their Kobweb folder
2 levels deep, and the message can now say:

```
We suggest typing:

$ cd site/app
$ kobweb run
```

Fixes #187