Change Kobweb CLI template filename
... from .kobweb/template.yaml to .kobweb-template.yaml

This change is important because soon we will allow specifying
Kobweb projects where the .kobweb folder is no longer at the root
level, but can be in a subdirectory. Therefore, we needed a way to
specify a different file that lives in the root of the project,
separate from the .kobweb folder.

Bug #187