Allow starting a Kobweb server from any folder
The idea being, if you can call `kobwebStart` from Gradle, perhaps
will a full Gradle path like `:site:kobwebStart`, then you have
enough information to know where the Kobweb root folder is.

Note that this doesn't change the behavior when calling `kobweb run`
via the binary. In that case, you still need to be in the correct
folder to use that call.