Expose some Gradle dependencies as API dependencies
When you *use* the Kobweb application or library plugin, the
core plugin is automatically inserted as a dependency for you.
The fact that there's even a core plugin is supposed to be
invisible to the user. (It's where KobwebBlock lives).

This change just does this for the code as well, in case anyone
writes their own Gradle plugin which references either the
application or library plugins as a dependency.