Add "slug" to pagecontext
I found myself needing this for analytics purpose in a project
I was working on, and the place I expected to find the current
slug value was in the page context.

It's true you can calculate this on your own using javascript
APIs, but by using the PageContext to get this, I could use it
as a key for a LaunchedEffect and be guaranteed to get notified
each time it changed.