Make kobweb.conf prod script required now
The prod script path was not initially added, so we kept it as a
nullable value for a while, but at this point, it should be safe
enough to expect people to have it set. Otherwise, they're going
to end up with a final JS file that's 2-3x larger than it should
be.