Fix issues with new ComponentStyle delegate methods
First, I was forgetting to make `colorMode` accessible in them.

Second, although it probably would never have happened in practice,
there was some ambiguity in the way I was generated variant names
and color-mode suffixed names.

Third, I added an extra "componentStyleBase" delegating method so
that it's quick to change between "componentStyle" and
"componentStyleBase" when developing.  Otherwise, it's identical to
"ComponentStyle.base"