Remove a case with the Link API being ambiguous
If you write Link("https://example.com"), that should work! But
currently the way Link is declared, Kobweb isn't sure which one to
choose, because it thinks you might want to use the one with a
custom render callback. This has been resolved by making the callback
required, not optional, so you either have this:

Link("...")

or

Link("...") { ... custom rendering ... }