Fix bug with pathless navigations leaking route prefix into interceptors
That is, something like `Link("?a=b")` or `Link("#test")` would
end up causing the route prefix, if set, to sneak into the path
passed into an interceptor, due to invalid assumptions.

This has been resolved by moving the prefix stripping logic a
few steps down in the algorithm that processed the passed in path.