Flip "show upgrade" message default from true to false
Probably we want to be more careful about when to show it vs.
when to hide it. If we don't show it when we could have,
that's a missed opportunity to announce an upgrade, sure, but
if we show it when we shouldn't have, that can be a really
frustrating and even potentially destructive experience.