Remove the "hideImmediately" logic from the StayOpenStrategy class
After thinking a bit more about it, this probably just complicates
things for no added benefit.

I was originally worried that a "hideDelay" of 500ms plus a
ManualStayOpenStrategy which got set to false when a user, say,
clicked on a button, it just didn't feel right. However, in
practice, a user should just set a really short (or 0ms) hideDelay
when used in conjunction with such a strategy.