Add handling for case where "hide" transition doesn't happen
It should be a pretty rare, but not impossible, case, where a
popup hide transition won't fire because the element is already
hidden. If this happens, our state machine never moves from the
"hiding" state to the "finished hiding" state.

We handle this by checking for it directly now, and short-circuit
that state machine transition if we detect this case.