Shuffle PageContext logic around to avoid extra recompositions
This is accomplished by the introduction of a "route" subproperty
in PageContext.

So this:

  ctx.router

doesn't change, but this:

  ctx.params.get("id")

becomes this:

  ctx.route.params.get("id")

This change is important because it means only users who query
the page route property will get hit with recompositions if things
change, while people who just want access to the router will not
experience any.