Add an optional parameter to renderActivePage which can wrap the page composable
Currently, calling `rememberPageContext` within an overlay (e.g. a
popup) will crash because the contents of the overlay are composed
OUTSIDE of the scope where the page context is nulled out. This
could be caused by a `Link` inside a popup, because `Link` (as one
example) delegates to `rememberPageContext` under the hood.