Rename "StayOpenStrategy" to "KeepOpenStrategy"
"Stay" is intransitive, while "Keep" in transitive, and has more
of a feel of the user taking a direct action with their strategy.

I found that every time I went to experiment with this strategy,
my mind kept (haha) assuming that it would be called "keep open"
since I was making an active decision whether or not the popup
should be open or not, so I thought that was probably a sign I
should change the name.