Move "StyleVariable" to "compose-html-ext" module
and also add an HTMLElement.setVariable extension method.

This allows users to set variables on raw HTML elements. It will be
especially important in the near future as we will use variables
throughout Silk styles soon, and we'll need to hook those variables
up to be set on the "root" element.