Remove examples; change "helloworld" to "playground"
This is to avoid confusion since otherwise people might think
that my intention is for this location to become a growing
collection of examples for Kobweb. Rather, those belong as
templates! This folder is really more of a workdir where a
Kobweb dev can test out some ideas fairly quickly.