Simplify GridTrackSize.minmax
It is now a bit more awkward to use directly, but it is more
consistent with the rest of the GridTrackSize APIs.

For example, this:

GridTrackSize.minmax(0.px, 1.fr)

became this:

GridTrackSize.minmax(GridTrackSize(0.px), GridTrackSize(1.fr))

That's awful, yes, but more consistent with how everything else
works (`gridTemplateRows(GridTrackSize(0.px)...)` and `repeat`)

If you use the grid builder, however, you can enjoy the nicer
syntax there:

```
grid {
   cols { minmax(0.px, 1.fr) }
}
```

which is also consistent with the rest of the GridBuilder API.
In other words, the GridTrackSize class has been made more simple
but also more verbose, while GridBuilder is the layer where
complexity goes into to enable a more enjoyable API experience.