Refactor Kobweb streams directly into kobweb-core
Originally, I was worried about adding Kotlinx serialization as
a dependency into kobweb-core, as I worried that it might add
a bunch of extra bytes to a project. However, experimentation
shows I was probably overthinking it, and by putting streams
into kobweb-core itself, it makes the feature much easier for
someone to use.