Fix "rgba" implementation
For the longest time, Color.rgba(value: Int) was wrong.
It actually expected 0xAARRGGBB, not 0xRRGGBBAA.

To prevent breaking existing codebases that might be using
the old value, we deprecated the method and told people
to migrate to Color.argb, with the promise we'd eventually
fix the implementation one day.

That day is today.