Turn "ColorMode" into a CompositionLocal
Previously, color mode was global for the whole site. You
could query it using `getColorMode` and modify it using
`rememberColorMode`.

In contrast, you can now scope subsections of your site
to areas with their own overridden color mode, using the
`Surface` widget to accomplish this.

Something like this:

```
@Page
@Composable
fun HomePage() {
   ...

   Surface(colorModeOverride = ColorMode.DARK) {
     ... this area is always in dark mode ...
   }
}
```

Instead of `getColorMode` and `rememberColorMode`, you now
use `ColorMode.current` and `ColorMode.currentState` for
read-only and read-write access, respectfully. Deprecating
the name `rememberColorMode` is another nice benefit to
this change, as it, in retrospect, was the wrong name,
chosen when I was still a bit of a novice with some of the
more subtle design decisions used in Compose APIs.