Remove white and black color schemes
A user shouldn't be using these values when given
a chance to set a color scheme, because most widgets
expect different gradients of the same color, not
different alpha shades of a single color. Plus,
the alpha makes it hard for some widgets to know
what color text to render on top, because the final
result might be bright (if on a bright background)
or dark (if on a dark background)

Let's just remove these for now. They're unlike
the rest of the choices in the collection. We
can revisit adding them back later if they might
be useful somehow.