Change variant name to "SwitchInputVariant"
This used to be "SwitchCheckboxVariant" which is nice, but it is
a precedent that doesn't work for Checkboxes, which is going to do
a similar style organization as Switch.  There, it's going to be
called "CheckboxInputVariant" ("CheckboxChecboxVariant" doesn't
really work... ;) so we update the name here in advance to be
consistent.