Remove checkboxModifier and checkboxRef parameters (for now)
We may add these back, but for now, we're simplifying the API
to be more consistent with other widgets and to follow the
rules in the CONTRIBUTING.md document.

It can be very useful to access the underlying checkbox
element, for example to tag it with a label for a11y
reasons, which could be a strong decision to either
revisit this decision or at least revisit the Switch
API again later (perhaps adding a `label` parameter).
We'll have to see how common this is in practice.