Routing logic no longer tries to route to files
Before, routing to "/files/example.md" (e.g. from a Link) would try to
navigate to the file and then show a 404. However, this would prevent
users from actually opening / downloading the file.

Now, we only attempt to route if there is no file extension (or if the
file extension ends in ".html")