Remove convenience Image method
It actually causes issues for users as
`Image("/src", alt = "Description")`, which should be readable
code, causes an ambiguity compile error, which is a bad experience.

It was a mistake to add it in the first place, and it's not easy
to deprecate because of that same ambiguity issue. So let's just
rip the band-aid off and document it in the release notes.