Fix logic in MainTemplate around initializing Silk
To repro the problem:

1. `$ kobweb create app/empty`
2. Edit MyApp.kt, delete the empty "@InitSilk" placeholder method

Because we must now call initSilkWidgets (required after a major
refactoring a few versions back), the fact that this block was getting
skipped over if no init silk, keyframes, styles, or variants were
present causes the empty app to break. The empty app should NOT break
just because someone removes an empty InitSilk method.