Remove `Image` deprecation for now.
Since I realized we could rename the `alt` parameter to avoid the
compiler ambiguity, it seems like it could still be useful to
offer this override, especially since setting a description on an
image is so strongly encouraged.

The fact that the parameter name is different in this one case can
be a little weird, but I think of it how `Input` has a field called
`enabled` which is actually used to set the `disabled` attribute.