Fix invalid file extension check
The Kobweb router considers slugs with "." in them as
filenames and early aborts in that case, letting the
browser handle the request instead. However, "."
should at least be allowed in the middle of the path,
such as `docs/2.1/reference/SomeClass`.

Unfortunately, Kobweb would detect the previous path as
containing a file extenions "1/reference/SomeClass". That's
obviously a mistake!

Note that, in the future, we may tweak this logic a bit
further, since there may be valid cases for handling it
even when a slug has a "." in it. However, for now, at
least we close off an unintended bug.