README: Simplfy API stream section
When I wrote the docs originally, I tried to teach the
flexibility of API streams, as there are quite a few
ways to create them. But rereading it more recently,
I realize 99% of people will just use `rememberApiStream`,
so I decided to focus on explaining that up front.

I also tweaked the intial section, to break up a giant
wall of text.