Fix issue with fragments in links sometimes not working
The logic in Kobweb around handling fragments often
doesn't work, because when a new page is visited, the DOM
is torn down and rebuilt back up. If we try to scroll to
an element with some ID before that rebuild process is
finished, the request fails silently.

The biggest problem with writing logic like this is we
don't know...

1) if the target page even has an element matching the ID
2) how long it will take the target page to get rendered
3) when the page is considered done

so in the face of that uncertainty, we just try a handful
of times and then give up after a reasonable amount of
time has passed.

Fixes #422