Introduce WorkerFactory interface to the worker API
It felt wrong that the worker class was creating a
worker strategy only to get to its IOSerializers.

```
val ioSerializer = SomeWorkerStrategy().ioSerializer
...
```

Not only did it feel like an abuse of the single-
responsibility principle, but it could easily result
in very confusing behavior. For example, if you put
a println in your WorkerStrategy constructor, you
wouldn't expect it to print twice! But this was
happening because the worker itself instantiated
it (to get at the IOSerializer) and the worker
script itself created a copy (in main).

What was worse is that one of those
instantiations was happening on the
application thread! Which means if you
did: `init { postOutput(...)` you were
actually sending a message to the worker
from the application. That would never have
been intentional.

Now, we have a WorkerFactory class, which
separates the creation of IOSerializers
and WorkerStrategies. This adds a bit of
abstraction to the process of defining
worker logic, but it's not excessive and
is much safer from accidental mistakes:

```
// Before
internal class MyWorkerStrategy : WorkerStrategy<I, O> {
  override fun onInput(input: I) {
    postOutput(input.toOutput())
  }

  override val ioSerializer = ...
}

// After
internal class MyWorkerFactory : WorkerFactory<I, O> {
  override fun createStrategy(postOutput: (O) -> Unit) =
    WorkerStrategy<I> { input ->
      postOutput(input.toOutput())
    }
  }
  override fun createIOSerializer() = ...
}
```