Worker projects now put a worker.json file in their artifact
This way we can tell in the future which Gradle plugin (worker
or library) a Kobweb artifact came from.

Note that we also update the library plugin in this commit too,
which now creates a library.json file (symmetric with the
worker's file). And since we added that, we moved index
information out of the now deprecated `index.json` file and
into the `library.json` file instead.