Make "kobwebSyncServerPluginJars" use "dependsOn" for "kobwebStart"
... and stop hooking into `jsProcessResources`

We were hanging this task on process resources just because it kind
of feels like you're bundling relevant resources while a site is
getting built, but server plugins are not really resources, and
moreover, the site itself doesn't depend on them directly.

Instead, the running server is the only thing that cares about
whether plugins are found or not.

So, to be more technically accurate, we're using the "kobwebStart"
task, which is the one which spins up the server that ultimately
looks for those files.