Remove "toPackageName" from markdown route generation
That was only necessary for the Kotlin source files, not for the
route, and it was included here incorrectly.

In other words, the path "blog/2022/example.md" generates the
route "blog/2022/example" and Kotlin package "blog._2022", but
we accidentally copied the wrong code here and the "_2022"
leaked into the route.