Revert markdown tasks back to using the build script as an input
We are currently encountering some very hard to understand compile
errors generated by Gradle when any of the Gradle lambda properties
in the markdown block happens to capture something unserializable
in its closure.

In the future, there may be a better way than this, but for now,
since this is more robust, we are using this approach for the next
release.