Change Kobweb default route formatting to kebab-case
Before, the algorithm we used was a simple lowercase, which is fine for
single word url parts, but confusing for multi-word parts.

So now, "MultiWord.kt" generates the slug "multi-word" and not "multiword".

The logic for changing packages has also been updated to be consistent.

The package "corp.ourTeam.ourValues" will now generate the route
"/corp/our-team/our-values/"

In order to support legacy sites, however, router code has been updated to
allow old links to work as well, although this can be configured.

As part of this change, we were also able to simplify the some of the
route logic inside the markdown plugin.