Hardcode the "public/" path for now
This lets us delete some code and simplify other paths as a result.

This is part of a bigger change, to move the "pagesPackage",
"apiPackage", and "publicPath" Gradle properties out of the root
Kobweb block (as they don't belong there, especially since we now
have `app`, `server`, and `library` blocks).

While I was looking at the public path property, I realize we
hardcode "public/" in some places and allow it to be overridden
in others. The task for copying supplemental resources from
libraries is a great example.

It seems unlikely ANYONE is using this right now, so for now, we
just remove it. Maybe we can add it back again in the future, if
requested -- but we'll need to think about how to handle this in
libraries, where they can also specify their own resources, and they
too should have a say. But if they do, then the app has to respect
their setting and use it. This seems like a lot of complexity for
no requested benefit.