Remove default arguments to BreakpointSizes for rem values
Probably code that got checked in by accident a long time ago,
and it's a relatively unused portion of the codebase so it was
easy to miss.

Note that it should definitely NOT be default sm to 0.rem, but
also, for consistency, the other `BreakpointSizes` methods to
not include defaults either.