Add redirect support to Kobweb.
Defining a simple redirect looks like this:

```yaml
server:
  redirects:
    - from: '/old-page'
      to: '/new-page'
```

and, when defined, both the Kobweb client AND
Kobweb server codebases will respect them.

This feature also supports regexes:

```yaml
server:
  redirects:
    - from: "/socials/facebook/([^/]+)"
      to: "/socials/meta/$1"
    - from: "(/socials/meta)/about-facebook"
      to: "$1/about-meta"
```

Note: This feature is orthogonal to the legacy redirect feature
added previously. However, we did clean up the code for that
feature while working on this.