Fix resolve logic to return non-null only for pages
Originally, the route "a/b/c" would resolve if "c" was associated
with a `@Page` but "a/b" would not (unless also associated with a
page). This constraint broke recently but was just fixed.

Most users wouldn't notice, but what it did do was break the fact
that "a/b" would automatically suffix a slash (to "a/b/") if that
generated a valid route. Instead, because we aggressively matched
with the non-slash version after the original logic was broken,
we wouldn't realize we should try appending a slash.