Move "defaultRoot" property to `markdown` block.
I originally put the "defaultRoot" property in the handlers
block, likely since that block has a bunch of other properties
in it, but when I went to look for it recently I couldn't find it.
I expected to find it in the top-level markdown block.

In fact, it's more confusing because we introduced an "imports"
property that lives in the markdown block, and "defaultRoot" is
definitely a similarly scoped feature to that.

Honestly, I currently think it makes way more sense to keep the
handlers block just for overriding how markdown concepts are
"rendered" in Kotlin, not for general markdown config settings
(like "defaultRoot" is).

So I went ahead and moved "defaultRoot" to there, deprecated the
previous property, and added some documentation to the README about it.
(Thankfully, the feature wasn't really documented before, so it's
unlikely that almost anyone is using it).