Update logic around how querying CssStyle names works
After this commit...

* Any CssStyle declared as a property (e.g. `val MyStyle = CssStyle { ... }``)
  will have a name.
* `SilkTheme.registerCssStyle` now MUST take a name. It's no longer optional.
* The SilkTheme is now responsible for managing the names of all CssStyles, which
  can be queried as `SilkTheme.nameFor(MyStyle)`
* The "by" keyword is no longer supported for `CssStyle`. This is a relic of the
  past -- all hail our new and improved KSP future!
* Fixed the TODO around Silk variables not being set on the room element anymore
* Relevant playground code snippets have been updated and they work again.

I also tweaked the FrontendProcessor logic which was generating duplicate "registerCssStyle"
calls.

In this commit, we do NOT update similar logic for ComponentStyle and ComponentVariant
classes. However, we intend to make that change later. The plan is to move all the naming
logic (e.g. "MyButtonStyle" -> "my-button") to the KSP code.